pyproject.toml
[tool.poetry]
name = "amagicalchildhood"
version = "0.1.0"
description = ""
authors = ["Vijayeta Rangarajan <vijayetar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.11.1"
gunicorn = "^20.0.4"
whitenoise = "^5.1.0"
django-environ = "^0.4.5"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
